// app/jobs/updateAnimals/utils/types.ts
export type RawAnimal = {
  animal_id: number;
  animal_subid: string;
  animal_kind: string;
  animal_sex: string;
  animal_bodytype: string;
  animal_colour: string;
  animal_remark: string;
  animal_opendate: string;
  animal_status: string;
  album_file: string;
  shelter_address: string;
  shelter_tel: string;
};